implement map transitions
See JumpToMap.cs for usage.

The ability of the player to click to move and interact with objects
is moved into the ClickToInteract behaviour.

We separate out a HexGridManager class from the existing HexGrid
class. This manager persists throughout map changes, and implements
the LoadMap method that is the workhorse of map changes.

To perform a map change is relatively straightforward in this scheme:
- The HexGridManager destroys its child HexGrid game object (if any)
  thus destroying *its* children, i.e. the HexCells and NPCs.
- The HexGridManager instantiates a new HexGrid from a prefab.
- The HexGridManager calls the `Setup` method of the newly created
  HexGrid in order to create the new HexCells and NPCs.

Note that this process of loading a new map will also destroy the
player GameObject, since it is a child of the HexGrid!
Hence, once the new map is loaded, the player must be respawned, and
warped to the correct location.

This actually solves some nasty problems related to the HexGrid
reference inside the MapEntity component: this reference would become
stale if the player GameObject were not destroyed and recreated upon
changing maps.