refactor event system setup
Previously, the EventManager would find all its relevant children's
components using GetComponentInChildNotNull, at runtime in Awake. This
is somewhat irritating because it requires a certain ordering of Awake
calls to work correctly (sometimes). Since these references are
unchanging, its simpler to hook up the references in the inspector,
which is what this commit introduces.

Also, many fields of the EventManager are renamed for clarity, since
the event manager will soon include the inventory system.
ScriptableObjects that used to live under Scripts/Events have been
moved to the ScriptableObjects directory. That directory will probably
need to be split up later.

This commit also introduces the inventory system part of the
hierarchy, but doesn't implement any behaviour for it.