MapEditor: eliminate cellUnderCursor member
Instead, we just pass the cell down to the places we need it. This
could definitely be refactored further.
The reason I did this was that cellUnderCursor was getting
mysteriously nulled out (according to some judiciously placed
Debug.Log statements) so instead, fuck members, and let's just use
nice parameters.
Thinking about it now, it might have been due to serialization, since
the member was private and not marked as SerializeField.