introduce expression language; improve serialization
Expression language:
* Motivation: need a flexible way to express conditions in scripted
  events, e.g. "player has >5 sticks in their inventory" or
  "progress in quest Q is 4".
* Solution:
  - IExpression<T> represents a computation that computes a
    result of type `T`.
  - Introduce Condition : IScript that holds an IExpression<bool> and
    two IScripts: one if the bool is true, and one if the bool is
    false.
  - Following the refactoring that gave us SubtypeControl<T>,
    implementing an editor system for expressions was easy.

This gave rise to a new problem involving serialization.
Specifically, the default C# SurrogateSelector is too stupid.

Background: It uses a strict type equality check to decide what
surrogate to use. I believe that behind the scenes, it uses a
hashtable to associate types to their registered surrogates.
Previously, this wasn't too much of a problem: we could enumerate all
types deriving from UnityEngine.Object and register the same surrogate
for each of them.

However, this wouldn't work correctly for generics! Why?

Consider the class `Variable<T> : ScriptableObject, IVariable<T>`
that holds a value of type `T`.
Consider also a type `[Serializable] class ReadVariable<T> :
IExpression<T>` that reads the value of a `Variable<T>` field.
During surrogate registration, we enumerate all subtype of
UnityEngine.Object and find the type `Variable<>`. This is a *generic
type definition*; there's a hole for the type parameter.
We associate to `Variable<>` our `UnityObjectSurrogate`.
However, at runtime we don't actually ever *have* any value of type
`Variable<>`! We have values of type `Variable<T>` for some concrete
choice of `T`.
In the case of the condition system, we need to cook up an
`IExpression<bool>`, so if we choose the "read variable" subtype to
implement it, we get the concrete type `Variable<bool>` that needs to
be serialized.

And here's the crux of it: `Variable<bool>` is not equal to
`Variable<>`! Consequently, the C# SurrogateSelector will blow up and
tell us that `Variable<bool>` is not marked serializable.

The solution:
We introduce a new type
`AssignableSurrogateSelector : ISurrogateSelector` that uses
`IsAssignableFrom` to decide whether to use its one and only bundled
`ISerializationSurrogate`.
Then, our BinaryFormatter has this surrogate selector assigned to it.
Consequently, any time our formatter encounters a type `T` to
serialize, it will be checking whether `T` is assignable to
UnityEngine.Object.

And you know what?
`Variable<T>` is assignable to UnityEngine.Object
*for any choice of `T`*!

Fuck you Unity. I win again!