player won't spawn out of bounds
Solution was to introduce a notion of "runtime value" for variables.
When the variable is initialized, it sets its runtime value to its
fixed value, and during runtime only modifies the fixed value.
This way, when play mode ends, the runtime value is still visible in
the editor, but the fixed value has not been stomped.