shop in same map as city; repositioning; auto-save scripted events
The auto-save and undo use the same functionality in Unity
(Undo.RecordObject) but it doesn't really seem to work for undo.
In particular, it seems like Unity registers tons of undo entries that
seemingly do nothing.
It might be possible to just record once in ScriptedEventEditor but
then we lose the granularity about *what* changed. Maybe this doesn't
matter.

Repositioning is a feature added to the IPositioner interface, added
as a supertype for what used to be called InitialPosition and
IsoInitialPosition.
Instead, we now have IsoPositioner and HexPositioner.
These attach to appropriate PositionVariable<T> subtypes and monitor
events on them. Previously, positioners would fire once on Start in
order to put the player and NPCs in the right places, according to
their destination variables.
Now, it's possible to fire this positioning logic on request by using
PositionVariable<T>.RequestReposition(). IPositioner objects monitor
the associated event PositionVariable<T>.RepositionRequested and
perform their positioning logic again.

This gives us a more performant / flexible way of doing map
changes. Specifically, for the Quebec shop, rather than perform a
*full* map change to a separate IsoGrid, we can continue to exist in
the same grid, and just warp the player to a distant area that
contains the shop.
Going forward, we might actually have the shop prefab overlap the shop
area in the map, and grey out the map and enable to shop.
This way, warping wouldn't even be required.