reorganize Misc directory and improve FadeDirection
I implement an eliminator for FadeDirection objects, allowing
convenient case analysis. Using this, I reimplement the existing
FadeDirection.GetInitialValueAndMultiplier using the eliminator, and I
implement the FadeDirection.Inverse method.

This is necessary because the FadeScreen event item specifies whether
the *screen* fades in or out, which is the *opposite* of whether the
fade-to-black panel is fading in or out.