add IsoPositionAnchor behaviour
This behaviour is used so that we can avoid hard-coding positions in
event scripts. Instead, we can create anchor objects in the scene.

Anchor objects have the IsoPositionAnchor behaviour which sets a
position variable specific to the anchor when the object's Start is
invoked. This variable can then be read later when needing to perform
warps. Consequently, it suffices to move the object in the scene to
change the warp destination. The upshot is that objects *containing*
the anchor can be moved in the scene and the warp destinations will
semantically be preserved.