fix exception in WriteVariable
WriteVariable could raise an exception because it enumerables all
interfaces implemented by a ScriptableObject in order to determine
what type of variable is being assigned to. Specifically, we want to
know what T is in IVariable<T>, so that we can form the generic type
IExpression<T> to be assigned to the variable.

We were simply enumerating all interfaces and invoking
GetGenericTypeDefinition, which fails if we get an interface that
isn't a generic type.

The resolution is to check first that the interface is a generic type
(with IsGenericType) before trying to obtain its generic type
definition.